Createe xy_to_gps_conv.py
This python script takes in the GPS coordinate of the SW most controller, the bearing in degrees to the North most controller, and the distances between each controller. It will output the accurate GPS locations of the other controllers based on the input GPS coordinate to ensure an accurate coordinate system for location processing.